+++ src/domain/article/schema/articleSchema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { createdAt } from '@/common/schema'
+import { createdAt } from '@/common/schemas'
 
 export const articleSchema = z.object({
   articleId: z.bigint(),
